Add require_modification to Sequel::Model, for checking that model instance updating and deleting affects a single row
Previously, Sequel would not check that model instance update and
delete queries actually modified a single row, which led to the
following behavior:

  DB.create_table(:as){primary_key :id}
  class A < Sequel::Model; end
  a = A.create
  a.delete
  a.save # no error!
  a.delete # no error!

This commit adds a new global, class, and instance flag to
Sequel::Model that makes the model instance update and delete
methods check that a single row was modified:

  Sequel::Model.require_modification = false # global
  Album.require_modification = true # class
  album.require_modification = false # instance

This appears to work fine on all tested database adapters except
ADO, native MySQL, and the mysql do subadapter.  Currently, the
global default is set to true, but I will be changing it in the
future to check with the database or dataset object to see if
it is supported.