Add Dataset#provides_accurate_rows_matched?, for seeing if update and delete are likely to return correct numbers
Some adapters don't have support for returning the correct number of
matching rows (such as ADO).  Others, such as MySQL using the native
and do adapters, may return a number that isn't the same as the
number of matching rows.  This flag allows you to check if the
dataset supports this and act accordingly.

The Sequel::Model code uses this flag to determine whether to
set require_modification for the model class if you haven't set a
global default.