Move Dataset code into appropriate files, add RDoc sections
Before, the Sequel::Dataset RDoc page was pretty messy, as just
listed the methods in alphabetical order without any sort of
grouping.

I just recently discovered RDoc sections, which allow you to
group related methods into different sections on the RDoc page,
which allows the user to easily focus on only the methods they
are probably interested in.

This commit adds sections to all of the Dataset code, separating
the methods the users probably don't care about with the methods
they probably do care about.  Unfortunately, I haven't yet
discovered how to to set the order of the sections, so that the
most important sections are first, but hopefully that can be
added later.  Even just this is a big improvement.

To make sure each method was in the correct section, I moved all
of the methods out of dataset.rb and dataset/convenience.rb and
placed them in the appropriate file.  I also added dataset/misc.rb
and dataset/mutation.rb to group the related methods together.
I discovered some more methods in the wrong section, and moved
those to the appropriate section.