Make RCTE plugin work when a :conditions option is used
The conditions option has to be handled differently than normal associations. It must be respected in the recursive CTE otherwise we may select orphaned records.
In order for a condition to work, it must reference qualified columns because the recursive select statement effectively selects from the source table twice. And in order for qualified conditions to work, I had to change the last select statement of the generated sql to alias 't' to the natural table name.