Support returning the number of rows updated/deleted on MSSQL when using the ADO adapter with an explicit :provider
I'd like to be able to do this for ADO in general, but it appears as
though ADO uses pass by reference with an integer to set the number
of rows affected by the query, which isn't supported by ruby.  As I
don't know how to work around this, I instead added some code to
the MSSQL ADO subadapter to use a second query to get the number of
rows modified.  However, since I believe this requires the same
connection to be used, it's only enabled when an explicit :provider
is given, since the default provider uses a new connection for
each query.  Reduces the test suite unhandled failures from 24 to 8
when using the SQLNCLI10 provider.