Minor fix to the postgres specs
Way back when, the postgres adapter had a default schema of public.
This caused problems in certain cases, which is why it currently
uses the Sequel default of nil.  However, the adapter spec still
set it back to :public in an after filter.  This resulted in the
default schema being used for all of the integration tests.

Since Sequel's default is the nil default_schema, the integration
tests should probably be run with that.