Fix MSSQL temporary table creation, but watch out as it changes the table name (Fixes #299)
This isn't a good solution, but there's not really a better one.
This does allow you to create temporary tables on MSSQL using
the :temp option to Database#create_table, but on MSSQL that
requires changing the table name to be prefaced with a #,
and always referencing the table name with the preceding #.

I have a feeling this will still break most Sequel code
that deals with temporary tables and works fine on other
databases, but it is better than the current sitution of
generating invalid SQL.  To fix it properly would require
adding a temporary_table_name(name) method to Database,
and telling people writing database independent temporary
table code to always use that.  I may do that in the future
if it is requested.