Make the integer_migrations directory valid
Coming changes will make it so you can't skip versions for
integer migrations, so remove the 5th file in the
integer_migrations directory.  However, there is an implicit
spec for file names where the first character after the version
number and underscore is a number, so make that change to the
alt_integer_migrations directory.