Save migration version after each migration, instead of after all migrations
Previously, Sequel didn't save a migration version until after all
migrations ran.  That meant that if you used the migrator to apply
multiple migrations at once, and one of the later migrations raised
an exception, Sequel would leave the migration version number as it
was originally.  When you fixed the migration that raised the
exception, and reran the migrator, it would try to reapply
migrations that had already run successfully, resulting in negative
consequences.

This changes the migrator to set the migration version after each
migration, which should fix the issue.

Related to this fix is now Sequel will create the schema_info table
if it doesn't exist before creating any migrations, and it will
always try to populate the table with a single row.  Also, if the
schema_info table has multiple rows, Sequel will raise an error.