Switch Sequel.migration DSL to use a new SimpleMigration class
The SimpleMigration class is a lot simpler than the Migration
class.  Instead of using subclasses for each migration, it uses
normal instances.  The up and down blocks inside the
Sequel.migration DSL are just assigned to attributes, and when
the migration is applied, the blocks are just instance_evaled
on the database, so you don't need method_missing.