Add sharding plugin
The sharding plugin makes Sequel::Model objects work well with Sequel's
sharding support.  It allows you to create objects on specific shards.
It remembers which shard you used to retrieve an object, and saves it
back to that shard.  It also works with associations, so associated
objects and association modification methods only deal with the same
shard as the main object.