Minor doc change to migration extension
Now that the migration extension is included in the core
RDoc sections on the website, note that classes inside
the extension are related to the extension, so that if
users browsing the core RDoc see them, they are made
aware they they aren't available by default.

The reason for including the extension in the
main RDoc is for the links to be set up correctly.