Change the minimum version to use the TimestampMigrator
Before, it used the unix epoch timestamp for Jan 1, 2005.
While writing the migration guide, I realized it was
possible to use other timestamp formats, and since it
is unlikely anyone is going to have 20000101 integer
migration versions, and I doubt anyone would use a
timestamp version lower than that it seemed like a good
place to start.