No longer apply association options to join table dataset when removing all many_to_many associated objects
In hindsight, this was a poor decision to add, since the options
can add filters on rows that are in the associated model table
and not the join table, which the join table will not have
access too.  Assuming that filters affect the join table instead
of the associated model table is probably wrong in more cases than
it is right.