Always use the association's dataset when removing associated objects using a primary key value
Before, it would attempt to do a direct lookup using the associated
class for the many_to_many case, since it is probably faster.
However, it complicates usage with the sharding plugin, and rather
than add another internal method that can be overridden, it's easier
just to use the association's dataset.