Make Model::Errors#empty? handle attributes with empty error arrays
This fixes a bug that occurs if you do something like:

  errors.add(:column, "Error") unless errors[:column]

Because Errors#[] automatically creates an empty array for the key,
Errors#empty? was returning false (since that's what Hash#empty?
does).  However, this is incorrect, as there aren't any errors.

Fix this by defining an empty? method that checks that count == 0.