Modify how before and after hooks are handled in some places
Sequel wasn't always consistent with how before and after hooks
were handled previously.  In some cases when overriding, before
hooks called super before the rest of the method, in some cases
after, and when they called it before the method, they didn't
always check for a return value of false.

To be consistent, change the documentation and implementation
to state that when a before hook is overridden, it should call
super at the end of the method, and if an after hook is
overridden, it should call super at the beginning of the method.
That way before hooks are always called in reverse order of
inclusion, and after hooks are always called in order of
inclusion.  This also makes it so you don't have to remember to
use return false if super == false, since that's not required
if the last expression in the method is super.

The implementation changes affect when super gets called, which
could potentially affect application behavior if you are
expecting certain before hooks to cancel other before hooks by
returning false.

There are some other related changes as well.  BEFORE_HOOKS and
AFTER_HOOKS constants got added to Sequel::Model, to make it
to differentiate between them.  The hook_class_methods and
instance_hooks plugins were refactored slightly.

The hook_class_methods plugin no longer skips later after
hooks if a previous after hook returns false, and the run_hooks
private instance method is now gone.

The instance_hooks plugin replaced the run_instance_hooks
private instance method with run_before_instance_hooks and
run_after_instance_hooks.