Make Model::Errors#on always return nil if there are no errors on that attribute
Previously, Errors#on returned nil if the column has not in the hash, but
if the column was in the hash, it returned the value of the column, even
if the value was empty.  This could happen with the following code:

  errors # => {}
  errors[:name]
  errors # => {:name=>[]}
  errors.on(:name)

Previously, that errors.on call would have returned [], but now it returns
nil.  This makes it consistent with the recent change to have empty? return
false if there are empty area values in the hash.