Fix leak of JDBC Statement objects when using transactions on JDBC on databases that support savepoints
This actually does some significant refactoring.  Instead of dealing
with a single statement object for the transaction, act like all
the other adapters and use a connection object, with a new statement
object created for every BEGIN/ROLLBACK/COMMIT.

This removes the private transaction_statement_object from the
JDBC adapter.