Make sure certain alter table operations clear the schema correctly on MySQL (Fixes #301)
This changes the order of operations so that the schema cache is
cleared after the alter table operations instead of before.  So if
any code needs to check the existing schema inside the alter table
operation, it won't cache a previous schema entry that is then
invalid after the alter table call.