Correctly handle more than 2 hierarchy levels in the single table inheritance plugin
Previously, Sequel did not handle more than 2 hierarchy levels in
the single table inheritance plugin.  If you were using more than
two levels, the documentation told you that you had to set up your
own custom filter.

With this change, you can now have as many class hierarchy levels
as you want, and Sequel will handle things automatically.

There shouldn't be any backwards compatibility problems, but this
does switch subclasses to use IN instead of = in their filter. That
may have performance implications, but any decent database is going
to treat an IN with one element the same as an =, so there should
not be any performance regression in the 2 level hierarchy case.