Add a Sequel::NotImplemented exception class, use instead of NotImplementedError
This modifies some default methods that are overwritten in all
adapters to use a Sequel::NotImplemented exception class.  It
also adds the tables and indexes method that are implemented by
many of the adapters to the main Database class, and has them
raise Sequel::NotImplemented.

This also modifies the indexes methods in some of the adapters
to take an options hash, even though they don't use it, for
consistency with the JDBC adapter and now the default method
signature.