Move to using explicit methods instead of aliases for methods that can be overridden in subclasses
Aliasing a method in ruby does not create a method that calls
another method, it creates a method with the same body as another
method.  Sometimes that is what you want, but in Sequel we often
want a method that calls the original method, so that if the
original method is overridden in a subclass, calling the alias
calls the overridden method instead of the default method.

This changes Sequel to use explicit methods instead of aliases
for methods that could be overridden in subclasses.