Only remove primary key values when saving all columns if they have not been modified
Before, Sequel unconditionally removed primary key columns when
saving existing objects in all cases.  However, if you pass the
primary key column directly to save, it makes sense for Sequel
to include it in the update statement.  Likewise, it makes
sense for Sequel to include it if you use save_changes and it
has been changed.

Sequel used to always use all columns in a regular save, but
some databases such as MSSQL have problems with that, as they
don't allow you to modify the value of an identity column,
even to set it to the value it already has.  That's the reason
Sequel started removing primary keys from update statements in
the first place.  Now, Sequel only removes primary key columns
from the update values if the columns haven't been modified,
which should fix the issue.

Note that this does not mean you can update primary key values
and save objects and have things work correctly.  It's almost
enough, but to do it properly requires saving existing values
of the primary keys when loading the object, which slows the
loading process down by a factor of 3:

h = {:a=>1}
puts Benchmark.measure{100000.times{b.load(h)}}
=>  0.540000   0.030000   0.570000 (  0.574908)
puts Benchmark.measure{100000.times{c.load(h)}}
=>  1.650000   0.130000   1.780000 (  1.787350)

That's definitely an unacceptable slow down, so that will have
to go into a plugin.