Add a UpdatePrimaryKey plugin that allows you to update the primary key of a model object
This plugin just caches a copy of the primary key values when the
object is loaded from the database, and uses the cached copy
until after the object is saved, so that if you modify the primary
key value, it will use the previous value in the UPDATE WHERE
clause, and the new value in the SET clause.

This causes a substantial performance hit when loading model
objects, though the model load time is is generally small compared
to time spent in the lower levels of the stack.