Add an AssociationPks plugin that adds association_pks and association_pks= methods for *_to_many associations
These methods allow you to get the primary keys for all associated
objects, and to modify which objects are associated by passing an
array of keys to the setter.  The setter will add entries from the
array that are not already associated, and remove already associated
entries that are not in the array.

This is a fairly low level plugin that operates at the dataset
level and does not instantiate model objects.  This is faster but
it means that callbacks will not be called.  Also, this only
works with single primary keys, not composite primary keys.