Add ability to warn when skipping plugin/extension tests, using SKIPPED_TEST_WARN environment variable
When the SKIPPED_TEST_WARN environment variable is present,
Sequel will warn when it skips a test suite that it can't run
because the underlying library isn't available.

This also changes the specs so that you don't need the json
gem installed in order to run the plugin specs.  Previously,
it was required.