Use a better error message if a plugin can't be loaded
Before, it would only give you the error message for the
fallback case.  Now, it gives you the error message for
both the original case and the fallback case.  This helps
for extensions like xml_serializer that depend on other
libraries.  Before, the error message you received said
that it couldn't load sequel_xml_serializer, instead of
that it couldn't load nokogiri.