Add a Tree plugin for treating model objects as being part of a tree
This is based on the sequel_tree plugin originally part of
sequel_plus.  It provides similar features to rcte_tree, but
works on databases that don't support recursive common table
expressions.  Additionally, it offers features not previously
supported by rcte_tree, such as tree roots and siblings.

rcte_tree now loads the tree plugin, so you have all of the
advantages of the tree plugin when using rcte_tree.  Because
of ruby's method lookup process, a separate module is used for
the rcte_tree methods, as when they were loaded in the standard
way, the tree plugin instance methods overrode the rcte_plugin
association methods.