Cache prepared statements when using the native sqlite driver, improving performance
Previously, when using the native SQLite adapter Sequel would treat
prepared statements exactly the same as bound variables, by just
executing the prepared statement SQL with the bound variables. With
this commit, Sequel will create and cache a prepared statement
object on a per connection basis, and reuse the prepared statement
object if you execute the prepared statement again with different
bound variables.  So instead of having to parse the SQL given every
time the prepared statement is executed, it only needs to parse it
once when preparing the query, and then can just execute it with
new bound variables.

This also refactors parts of the native SQL adapter to make things
simpler.