Support transaction isolation levels on PostgreSQL, MySQL, and MSSQL
Transaction isolation levels are a way to specify how visible
outside changes are to a running transaction.  Sequel previously
did not have the ability to specify which transaction isolation
level to use, which means the database's default was always used.

This commit adds support for an :isolation option to
Database#transaction, which will modify the isolation level to
use for the transaction.  It also adds a
transaction_isolation_level Database attr_accessor, which specifies
the default transaction isolation level if one is not given via
the :isolation option.  On MSSQL, it's recommended that if you
ever use the :isolation option, you also set the
transaction_isolation_level attribute, as MSSQL transaction
isolation levels are per connection, not transaction.

The four transaction isolation levels supported by PostgreSQL,
MySQL, and MSSQL are :uncommitted, :committed, :repeatable,
and :serializable, for READ UNCOMMITTED, READ COMMITTED,
REPEATABLE READ, and SERIALIZABLE.

I looked into support for Oracle, Firebird, SQLite, and H2,
but while all of them have similar options, I don't test on
Oracle or Firebird, and SQLite and H2 transaction isolation
level support is substantially different from the SQL
standard to support using a separate API (which may be added if
requested).