Support bitwise operators on more databases
This commit expands the bitwise operator support to more databases.
In some cases where bitwise operators are not supported directly,
they are emulated using database functions. Specifically, the
following adapters now support the following bitwise operators:

* H2: &, |, ^, <<, >>
* MSSQL: <<, >>
* MySQL: ~

Previously, ~ on MySQL did work, but it returned an unsigned
integer.  It is now casted to return a signed integer, for better
compatibility with other databases.

Of the tested adapters, the only ones without full bitwise operator
support are H2 (which doesn't support ~) and SQLite (which doesn't
support ^).