Allow nil/NULL to be used as a CASE expression value
Previously, you could not use nil/NULL as a case expression:

  {1=>2}.case(0, nil)
  # CASE WHEN 1 THEN 2 ELSE 0 END

With this commit, if you provide a 2nd argument to case, the
expression is always used, even if it is nil, so you get:

  # CASE NULL WHEN 1 THEN 2 ELSE 0 END

This behavior makes more sense when you consider the following
scenario:

  {1=>2}.case(0, team_id)

If team_id is nil, you want to use NULL, because without the
NULL expression, the result would be 2 instead of 0 (or an
exception would be raised if the database doesn't allow integers
to be used for booleans).

Implementing this is fairly easy using a similar hack to the one
used for Symbol#*, which allows you to recognized whether or not
an optional was given, even if it has the same value as the
default.

This breaks backwards compatibility if you were previously
passing nil as the 2nd argument to case and did not want to
use an expression value.