Allow Dataset#grep to take :all_patterns, :all_columns, and :case_insensitive options
This is based on mighub's work on an :all option for Dataset#grep.
The :all_patterns option makes it so that the grep must match all
patterns to at least one column, and the :all_columns option makes
it so that the grep must match all columns to any of the patterns.
Using both means all columns must match all patterns.

There's also support for a :case_insensitive option, which is just
passed to SQL::StringExpression.like.  Previously, you could
get the behavior by including an option hash as the last element in
the second argument array, but that no longer works.  Dataset#grep
now accepts an explicit options hash.