Fix Database#each_server when using a connection string to connect
Database#each_server takes the options from the given server and
passes them to Database.connect.  However, if a connection string
was used instead of an options hash, the :adapter option may not
have been filled in, which made Database.connect raise an
exception.  Fix this by explicitly populating the :adapter option
when using a connection string.