Don't consider unsaved nodes as root nodes
This is a judgment call, as I can see the argument for
having new nodes be considered as a root if they haven't
explicitly set the parent_id column.  However, I think
this is a better default.  For one, it better handles
default values, either set by the database or in a before
save hook.