Make class_table_inheritance plugin work with non integer primary keys on SQLite
It was broken before because SQLite returns the hidden integer primary
key on insert if the primary key is not an integer, and the plugin
did not check for an existing primary key entry (like the standard
insert code does).

Also change before_create to call super last, per Sequel's convention.