Handle the case where both an implicit and explicit alias is given to join_table
Instead of attempting to use both aliases and yielding invalid SQL,
prefer the explicit alias.

This adds the helper method split_alias, which is much like the
unaliased_identifier method, except it returns an array that includes
the alias.

This appears to work with eager_graph, though it doesn't have much of
an effect as eager_graph uses a subselect for an aliased join.