Cache setter methods in the model for a big speed boost
Whenever a mass assignment is made, via .new/.create/#set/#update,
a large part of the time was taken to determine which setter
methods to allow.  For the common case, this can be cached in the
model class, which yields a very nice performance improvement:

* .create: 1.2x
* .new: 8.5x
* #set: 10.4x
* #update: 1.4x

Unfortunately, maintaining this code has a maintenance cost, as
all the class methods that could affect the cache need to clear it,
and you have to use method_added and singleton_method_added to
check for someone adding their own methods.  Kind of ugly, but
definitely worth it for the performance benefit.

This is not full proof, as I didn't add method_removed or
method_undefined hooks.  If you remove or undefine a setter method
and then try to assign to it, you'll still get an exception, though
it will probably be a NoMethodError instead of a Sequel::Error. I
expect that use case is infrequent enough that there shouldn't be
built in support for it.  Any affected users can use:

  MyModel.instance_variable_set(:@setter_methods, nil)

to clear the cache manually.