Make class MyModel < Sequel::Model(DB[:table]) reload safe
The reason this didn't work before is because Dataset did not
implement an appropriate hash or == method.  This commit does
that.  Both == and hash check the database, options, and SQL
for the dataset.  I was originally planning on having them
check the row_proc, but that wouldn't work correctly with the
existing implementation of Sequel::Model().