Be less picky on some integration tests
For tests on Date, as long as they have the same string format,
the represent the same date.  This fixes issues with adapters
that return dates with a local time offset.

In the timestamps plugin test, handle both Time and DateTime
objects, so that adapters that always use DateTime work.