Note that the identity_map plugin is not compatible with standard eager loading of many_*_many associations
This is because the same database row in the associated table can get
loaded multiple times with separate left_key_alias values, one for each
row in the current table that it is associated with. Without the
identity_map, this results in separate instances with distinct
left_key_alias values, which the eager loading code uses to associate it
to the primary model object. When the identity_map is used, only a
single instance is created with the first left_key_alias value, which
leads to incorrect results.