Add :replace option to Database#create_language in the postgresql adapter
The :replace option adds OR REPLACE to the SQL used if the server
version is at least 9.0.  On previous versions, the option is
ignored since earlier versions don't support OR REPLACE.

Note that the Sequel tests uninstall the plpgsql language handler,
but as you shouldn't be sharing the Sequel test database with
anything else, that's unlikely to cause problems.