A couple of changes to the swift adapter
This fixes a couple of issues with the swift adapter. On PostgreSQL,
it would log some statements twice.  On MySQL, it would use a new
server-side prepared statement for every SELECT, eventually running
out of server-side prepared statements and crashing.

On the PostgreSQL adapter, execute insert was not catching swift
exceptions and reraising them as DatabaseErrors, and on MySQL it
wasn't explicitly finishing the result object, so memory wouldn't
be freed until garbage collection was run.

With these changes and recent updates to swift/dbic++, the swift
adapter passes the Sequel specs for both PostgreSQL and MySQL, and
the swift benchmarks can be successfully run on the Sequel swift
adapter.