Fix Database#each_server when used with jdbc or do connection strings without separate :adapter option
Database#each_server calls Sequel.connect with the current Database
instance's options.  Unfortunately, for jdbc and do connection
strings, the connection string itself is not parsed into options,
it is passed directly to the underlying adapter.  Because of this,
the :adapter option is not filled in, and when #each_server calls
connect with the options hash, no :adapter option is present and
you got an error.

You could previously work around this using the :adapter option
specificially, but that was an ugly fix.  This makes all Database
instances store an :adapter_class option that points to the
adapter class currently being used.  That way, when each_server
is called, the correct adapter class will always be used.