Use a default wait_timeout that's allowed on Windows for the mysql and mysql2 adapters (Fixes #314)
Another MySQL stupidity rears its ugly head.  For some reason, MySQL
has different allowed value ranges for wait_timeout depending on the
server OS used.  The previous default value that Sequel used was
above the limit allowed on Windows.  This lowers the value to the
maximum allowed on Windows, for both the mysql and mysql2 adapters.