Do manual type_translation in the SQLite adapter
sqlite3-ruby's internal support for type translation is going away
in 2.0.0, so switch to using a proc system in the sqlite adapter
similar to the one used in the MySQL and PostgreSQL adapters, though
using strings as keys instead of integer type ids.

This works almost perfectly, but there is a minor backwards
compatibility issue.  Previously, Sequel assumed a timestamp
or datetime field that was a single integer was a unix epoch
integer datetime.  That complicated things and I don't think
anyone uses it, so it's no longer supported.