Add to_dot extension
This adds a Sequel::Dataset#to_dot method.  The to_dot
method returns a string that can be processed by graphviz's dot
program in order to get a visualization of the dataset.  Basically,
it shows a version of the dataset's abstract syntax tree.

Idea stolen from Aaron Patterson's similar method in ARel 2.