The instance_hooks plugin now only removes hook pairs that have been run successfully
Previously, any successful save would remove all instance hooks.
That usually didn't cause a problem, but if a user added a
before_destroy hook to an object, saved the object, and then
destroyed the object, the before_destroy hook would not be
run, since it was removed when the object was saved.

This commit changes things so that only related instance hooks
are removed.  This is simplest and I think the most logical
behavior.