When emulating schema methods such as drop_column on SQLite, recreate applicable indexes on the recreated table
I found this bug while working on the reversible migrations support.
Basically, on SQLite, most schema methods that alter a table, other
than adding a column, require dropping the table and recreating it.
In order to support a method like drop_column, internally Sequel
will drop the table and create a new table with the column dropped.
Before, indexes weren't copied over from the old table to the new
table.  Now, Sequel will parse the indexes from the old table,
and recreate the indexes on the new table as long as all columns
are still available.